@@ -23,9 +23,12 @@ class TextArtifact(BaseArtifact):
     Methods:
         __add__(self, other: BaseArtifact) -> TextArtifact: Concatenates the text value of the artifact with another artifact.
         __bool__(self) -> bool: Checks if the text value of the artifact is non-empty.
-        generate_embedding(self, driver: BaseEmbeddingModel) -> Optional[list[float]]: Generates the embedding of the text artifact using a given embedding model.
-        token_count(self, tokenizer: BaseTokenizer) -> int: Counts the number of tokens in the text artifact using a given tokenizer.
-        to_bytes(self) -> bytes: Converts the text value of the artifact to bytes using the specified encoding and error handler.
+        generate_embedding(self, driver: BaseEmbeddingModel) -> Optional[list[float]]:
+        Generates the embedding of the text artifact using a given embedding model.
+        token_count(self, tokenizer: BaseTokenizer) -> int:
+        Counts the number of tokens in the text artifact using a given tokenizer.
+        to_bytes(self) -> bytes:
+        Converts the text value of the artifact to bytes using the specified encoding and error handler.
     """
 
     value: str
