@@ -1,6 +1,6 @@
 from swarms import Agent, OpenAIChat
 
-## Initialize the workflow
+# Initialize the workflow
 agent = Agent(
     llm=OpenAIChat(),
     max_loops=1,
