@@ -1,6 +1,7 @@
 from typing import Any, Dict, List, Optional, Union
 
-import openai
+from openai import OpenAI
+
 import requests
 from pydantic import BaseModel, validator
 from tenacity import (
@@ -147,6 +148,7 @@ class OpenAIFunctionCaller:
         self.user = user
         self.messages = messages if messages is not None else []
         self.timeout_sec = timeout_sec
+        self.client = OpenAI(api_key=self.openai_api_key)
 
     def add_message(self, role: str, content: str):
         self.messages.append({"role": role, "content": content})
@@ -163,7 +165,7 @@ class OpenAIFunctionCaller:
     ):
         headers = {
             "Content-Type": "application/json",
-            "Authorization": "Bearer " + openai.api_key,
+            "Authorization": "Bearer " + self.openai_api_key,
         }
         json_data = {"model": self.model, "messages": messages}
         if tools is not None:
@@ -235,7 +237,7 @@ class OpenAIFunctionCaller:
                 )
 
     def call(self, task: str, *args, **kwargs) -> Dict:
-        return openai.Completion.create(
+        return self.client.completions.create(
             engine=self.model,
             prompt=task,
             max_tokens=self.max_tokens,
