@@ -37,9 +37,6 @@ from tenacity import (
     stop_after_attempt,
     wait_exponential,
 )
-
-logger = logging.getLogger(__name__)
-
 from importlib.metadata import version
 
 from packaging.version import parse
@@ -177,11 +174,11 @@ def _create_retry_decorator(
     import openai
 
     errors = [
-        openai.error.Timeout,
-        openai.error.APIError,
-        openai.error.APIConnectionError,
-        openai.error.RateLimitError,
-        openai.error.ServiceUnavailableError,
+        openai.Timeout,
+        openai.APIError,
+        openai.APIConnectionError,
+        openai.RateLimitError,
+        openai.ServiceUnavailableError,
     ]
     return create_base_retry_decorator(
         error_types=errors,
@@ -352,7 +349,13 @@ class BaseOpenAI(BaseLLM):
         try:
             import openai
 
-            values["client"] = openai.Completion
+            values["client"] = openai.OpenAI(
+                api_key=values["openai_api_key"],
+                api_base=values["openai_api_base"] or None,
+                organization=values["openai_organization"] or None,
+                # TODO: Reenable this when openai package supports proxy
+                # proxy=values["openai_proxy"] or None,
+            )
         except ImportError:
             raise ImportError(
                 "Could not import openai python package. "
@@ -645,9 +648,11 @@ class BaseOpenAI(BaseLLM):
             "organization": self.openai_organization,
         }
         if self.openai_proxy:
-            import openai
+            pass
 
-            openai.proxy = {"http": self.openai_proxy, "https": self.openai_proxy}  # type: ignore[assignment]  # noqa: E501
+            # TODO: The 'openai.proxy' option isn't read in the client API. You will need to pass it when you instantiate the
+            # client, e.g. 'OpenAI(proxy={"http": self.openai_proxy, "https": self.openai_proxy})'
+            # openai.proxy = {"http": self.openai_proxy, "https": self.openai_proxy}  # type: ignore[assignment]  # noqa: E501
         return {**openai_creds, **self._default_params}
 
     @property
@@ -933,44 +938,36 @@ class OpenAIChat(BaseLLM):
     @root_validator()
     def validate_environment(cls, values: dict) -> dict:
         """Validate that api key and python package exists in environment."""
-        openai_api_key = get_from_dict_or_env(
-            values, "openai_api_key", "OPENAI_API_KEY"
-        )
-        openai_api_base = get_from_dict_or_env(
-            values,
-            "openai_api_base",
-            "OPENAI_API_BASE",
-            default="",
-        )
-        openai_proxy = get_from_dict_or_env(
-            values,
-            "openai_proxy",
-            "OPENAI_PROXY",
-            default="",
-        )
-        openai_organization = get_from_dict_or_env(
-            values,
-            "openai_organization",
-            "OPENAI_ORGANIZATION",
-            default="",
-        )
+        # openai_api_key = get_from_dict_or_env(
+        #     values, "openai_api_key", "OPENAI_API_KEY"
+        # )
+        # openai_api_base = get_from_dict_or_env(
+        #     values,
+        #     "openai_api_base",
+        #     "OPENAI_API_BASE",
+        #     default="",
+        # )
+        # openai_proxy = get_from_dict_or_env(
+        #     values,
+        #     "openai_proxy",
+        #     "OPENAI_PROXY",
+        #     default="",
+        # )
+        # openai_organization = get_from_dict_or_env(
+        #     values,
+        #     "openai_organization",
+        #     "OPENAI_ORGANIZATION",
+        #     default="",
+        # )
         try:
             import openai
-
-            openai.api_key = openai_api_key
-            if openai_api_base:
-                openai.api_base = openai_api_base
-            if openai_organization:
-                openai.organization = openai_organization
-            if openai_proxy:
-                openai.proxy = {"http": openai_proxy, "https": openai_proxy}  # type: ignore[assignment]  # noqa: E501
         except ImportError:
             raise ImportError(
                 "Could not import openai python package. "
                 "Please install it with `pip install openai`."
             )
         try:
-            values["client"] = openai.ChatCompletion
+            values["client"] = openai.OpenAI
         except AttributeError:
             raise ValueError(
                 "`openai` has no `ChatCompletion` attribute, this is"
