@@ -447,7 +447,7 @@ class Agent:
                     Name: {self.agent_name}
                     Description: {self.agent_description}
                     Standard Operating Procedure: {self.sop}
-                    System Prompt: {self.system_prompt} 
+                    System Prompt: {self.system_prompt}
                     Task: {task}
                     Max Loops: {self.max_loops}
                     Stopping Condition: {self.stopping_condition}
@@ -778,7 +778,7 @@ class Agent:
 
                 Follow this standard operating procedure (SOP) to complete tasks:
                 {self.sop}
-                
+
                 {history}
             """
             return agent_history_prompt
@@ -786,7 +786,7 @@ class Agent:
             system_prompt = self.system_prompt
             agent_history_prompt = f"""
                 System : {system_prompt}
-                
+
                 {history}
             """
             return agent_history_prompt
