@@ -85,7 +85,7 @@ def tools_prompt_prep(docs: str = None, scenarios: str = SCENARIOS):
     You will be provided with a list of APIs. These APIs will have a
     description and a list of parameters and return types for each tool. Your
     task involves creating varied, complex, and detailed user scenarios
-    that require to call API calls. You must select what api to call based on 
+    that require to call API calls. You must select what api to call based on
     the context of the task and the scenario.
 
     For instance, given the APIs: SearchHotels, BookHotel, CancelBooking,
@@ -116,14 +116,14 @@ def tools_prompt_prep(docs: str = None, scenarios: str = SCENARIOS):
     different combination of APIs for each scenario. All APIs must be used in
     at least one scenario. You can only use the APIs provided in the APIs
     section.
-    
+
     Note that API calls are not explicitly mentioned and their uses are
     included in parentheses. This behaviour should be mimicked in your
     response.
-    
-    Output the tool usage in a strict json format with the function name and input to 
+
+    Output the tool usage in a strict json format with the function name and input to
     the function. For example, Deliver your response in this format:
-    
+
     ‘‘‘
     {scenarios}
     ‘‘‘
