@@ -6,7 +6,9 @@ from swarms.structs.agent import Agent
 from swarms.structs.groupchat import GroupChat, GroupChatManager
 
 llm = OpenAIChat()
-llm2 = Anthropic()
+# llm2 = Anthropic()
+# TODO: Mock anthropic class
+llm2 = OpenAIChat()
 
 
 # Mock the OpenAI class for testing
